Nagios Host Status email Notification
Nagios built-in email notification tries to email all the status change
information. In our situation, since there are thousands of hosts in the
system, we could receive thousands of email every day, which is
annoying. This script analyzes the status of the hosts from Nagios, and
send emails to managers every week to show the hosts which are down and
the duration of the status.